requirements.txt

Django==4.1.5
django-cors-headers==3.13.0
django-jazzmin==2.6.0
djangorestframework==3.14.0
drf-yasg==1.21.4
mysqlclient==2.1.1
Pillow==9.4.0
requests==2.28.2
